+++ packages/api/models/reconciliation-reports.js
@@ -6,8 +6,6 @@ const { reconciliationReport } = require('../lib/schemas');
 class ReconciliationReport extends Manager {
   constructor() {
     super({
-      tableName: process.env.ReconciliationReportsTable,
-      tableHash: { name: 'name', type: 'S' },
       schema: reconciliationReport,
     });
   }