+++ tf-modules/cumulus/migration_helper_async_operation.tf
@@ -1,25 +1,3 @@
-
-module "reconciliation_report_migration_lambda" {
-  source = "../../lambdas/reconciliation-report-migration"
-
-  prefix              = var.prefix
-  system_bucket       = var.system_bucket
-
-  dynamo_tables = var.dynamo_tables
-
-  lambda_subnet_ids   = var.lambda_subnet_ids
-  lambda_timeouts     = var.lambda_timeouts
-  lambda_memory_sizes = var.lambda_memory_sizes
-
-  permissions_boundary_arn   = var.permissions_boundary_arn
-
-  rds_security_group_id                  = var.rds_security_group
-  rds_user_access_secret_arn             = var.rds_user_access_secret_arn
-
-  tags                = var.tags
-  vpc_id              = var.vpc_id
-}
-
 module "migration_helper_async_operation" {
   source = "../../lambdas/migration-helper-async-operation"
 
@@ -41,8 +19,6 @@ module "migration_helper_async_operation" {
 
   prefix                     = var.prefix
   permissions_boundary_arn   = var.permissions_boundary_arn
-
-  reconciliation_report_migration_function_arn = module.reconciliation_report_migration_lambda.lambda_arn
 
   rds_connection_timing_configuration    = var.rds_connection_timing_configuration
   rds_security_group_id                  = var.rds_security_group