+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [v18.4.1] 2024-09-21
+
+  - Fix async Operation migration to include Migration Count Report
+
 ## [v18.4.0] 2024-08-16
 
 ### Migration Notes